Move MultiName to a new module
