+++ serde_derive/src/internals/mod.rs
@@ -4,6 +4,7 @@ pub mod attr;
 mod case;
 mod check;
 mod ctxt;
+mod name;
 mod receiver;
 mod respan;
 mod symbol;
